+++ cmake/SearchForStuff.cmake
@@ -114,7 +114,7 @@ if (PKG_CONFIG_FOUND)
 
   #################################################
   # Find DART
-  find_package(DARTCore 4.3 QUIET)
+  find_package(DARTCore 4.3.3 QUIET)
   if (DARTCore_FOUND)
     message (STATUS "Looking for DARTCore - found")
     set (HAVE_DART TRUE)
@@ -433,8 +433,7 @@ if (NOT GDAL_FOUND)
 else ()
   message (STATUS "Looking for libgdal - found")
   set (HAVE_GDAL ON CACHE BOOL "HAVE GDAL" FORCE)
-  include_directories(${GDAL_INCLUDE_DIR})
-endif()
+endif ()
 
 ########################################
 # Find libusb